[VIM] Install and enable Limelight plugin
The Limelight plugin will change the colour of the text not in the
immediate vicinity of the cursor. Full details and an illustration can
be found here https://github.com/junegunn/limelight.vim

Because the Solarized theme is enabled, the ctermfg colour must be
manually specified for Limelight since it does not support this theme
out of the box.

Limelight must be enabled using the VimEnter auto command since the
plugins are not loaded until after the .vimrc has been sourced, so
placing the Limelight command straight in there will result in an
error.